Introduce Server connection and generic dialogs
-Added generic dialogs for a more robust dialog fragment implementation
on Android.
-Server connections are now actually occurring! Although I have not
actually be able to get this to work, not sure where the problem lies.
-Moved the barcode integration to maven.

Lots of work is still needed but I'm stopping here today. Hopefully I
can fix the server connection issues.

+++ app/src/main/java/com/android305/forgeessentialsremote/servers/ServerAddFragment.java
@@ -40,7 +40,7 @@ public class ServerAddFragment extends Fragment {
     private CheckBox autoConnect;
     private CheckBox defaultServer;
     private boolean edit = false;
-    private long id = 0;
+    private int id = 0;
     private String uuid = null;
 
     public ServerAddFragment() {
